Separate fusion and compilation
